refactor event manager references to worker
