+++ Cargo.toml
@@ -11,7 +11,7 @@ members = [
   "./crates/sb_os",
   "./crates/node",
   "./crates/cpu_timer",
-  "./crates/event_manager"
+  "./crates/event_worker"
 ]
 resolver = "2"
