+++ crates/base/src/commands.rs
@@ -6,7 +6,7 @@ pub async fn start_server(
     ip: &str,
     port: u16,
     main_service_path: String,
-    event_manager_path: Option<String>,
+    event_worker_path: Option<String>,
     import_map_path: Option<String>,
     no_module_cache: bool,
     callback_tx: Option<Sender<ServerCodes>>,
@@ -15,7 +15,7 @@ pub async fn start_server(
         ip,
         port,
         main_service_path,
-        event_manager_path,
+        event_worker_path,
         import_map_path,
         no_module_cache,
         callback_tx,
