+++ crates/base/src/server.rs
@@ -2,7 +2,7 @@ use crate::worker_ctx::{
     create_events_worker, create_user_worker_pool, create_worker, WorkerRequestMsg,
 };
 use anyhow::Error;
-use event_manager::events::WorkerEventWithMetadata;
+use event_worker::events::WorkerEventWithMetadata;
 use hyper::{server::conn::Http, service::Service, Body, Request, Response};
 use log::{debug, error, info};
 use sb_worker_context::essentials::{
