+++ crates/cli/src/main.rs
@@ -33,7 +33,7 @@ fn cli() -> Command {
                 .arg(arg!(--"main-service" <DIR> "Path to main service directory").default_value("examples/main"))
                 .arg(arg!(--"disable-module-cache" "Disable using module cache").default_value("false").value_parser(FalseyValueParser::new()))
                 .arg(arg!(--"import-map" <Path> "Path to import map file"))
-                .arg(arg!(--"event-manager" <Path> "Path to event manager directory"))
+                .arg(arg!(--"event-worker" <Path> "Path to event worker directory"))
         )
 }
 
@@ -78,14 +78,14 @@ fn main() -> Result<(), anyhow::Error> {
                     .get_one::<bool>("disable-module-cache")
                     .cloned()
                     .unwrap();
-                let event_service_manager_path =
-                    sub_matches.get_one::<String>("event-manager").cloned();
+                let event_service_worker_path =
+                    sub_matches.get_one::<String>("event-worker").cloned();
 
                 start_server(
                     ip.as_str(),
                     port,
                     main_service_path,
-                    event_service_manager_path,
+                    event_service_worker_path,
                     import_map_path,
                     no_module_cache,
                     None,
