+++ crates/event_worker/js_interceptors.rs
